test(parser): Move parser tests next to there definitions
