+++ crates/toml_edit/src/parser/key.rs
@@ -55,3 +55,28 @@ pub(crate) fn is_unquoted_char(c: u8) -> bool {
 
 // dot-sep   = ws %x2E ws  ; . Period
 const DOT_SEP: u8 = b'.';
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    use combine::stream::position::Stream;
+    use snapbox::assert_eq;
+
+    #[test]
+    fn keys() {
+        let cases = [
+            ("a", "a"),
+            (r#""hello\n ""#, "hello\n "),
+            (r#"'hello\n '"#, "hello\\n "),
+        ];
+
+        for (input, expected) in cases {
+            let parsed = simple_key().easy_parse(Stream::new(input.as_bytes()));
+            assert!(parsed.is_ok());
+            let ((.., k), rest) = parsed.unwrap();
+            assert_eq(k.as_str(), expected);
+            assert_eq!(rest.input.len(), 0);
+        }
+    }
+}
