+++ crates/toml_edit/src/parser/numbers.rs
@@ -194,3 +194,63 @@ pub(crate) const NAN: &[u8] = b"nan";
 parse!(nan() -> f64, {
     bytes(NAN).map(|_| f64::NAN)
 });
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    use crate::parser::*;
+    use combine::stream::position::Stream;
+
+    #[test]
+    fn integers() {
+        let cases = [
+            ("+99", 99),
+            ("42", 42),
+            ("0", 0),
+            ("-17", -17),
+            ("1_000", 1_000),
+            ("5_349_221", 5_349_221),
+            ("1_2_3_4_5", 1_2_3_4_5),
+            ("0xF", 15),
+            ("0o0_755", 493),
+            ("0b1_0_1", 5),
+            (&std::i64::MIN.to_string()[..], std::i64::MIN),
+            (&std::i64::MAX.to_string()[..], std::i64::MAX),
+        ];
+        for &(input, expected) in &cases {
+            let parsed = numbers::integer().easy_parse(Stream::new(input.as_bytes()));
+            parsed_eq!(parsed, expected);
+        }
+
+        let overflow = "1000000000000000000000000000000000";
+        let parsed = numbers::integer().easy_parse(Stream::new(overflow.as_bytes()));
+        assert!(parsed.is_err());
+    }
+
+    #[test]
+    fn floats() {
+        let cases = [
+            ("+1.0", 1.0),
+            ("3.1419", 3.1419),
+            ("-0.01", -0.01),
+            ("5e+22", 5e+22),
+            ("1e6", 1e6),
+            ("-2E-2", -2E-2),
+            ("6.626e-34", 6.626e-34),
+            ("9_224_617.445_991_228_313", 9_224_617.445_991_227),
+            ("-1.7976931348623157e+308", std::f64::MIN),
+            ("1.7976931348623157e+308", std::f64::MAX),
+            ("nan", f64::NAN),
+            ("+nan", f64::NAN),
+            ("-nan", f64::NAN),
+            ("inf", f64::INFINITY),
+            ("+inf", f64::INFINITY),
+            ("-inf", f64::NEG_INFINITY),
+            // ("1e+400", std::f64::INFINITY),
+        ];
+        for &(input, expected) in &cases {
+            parsed_float_eq!(input, expected);
+        }
+    }
+}
